resolved errors with adding chapters with duplicated number; implemented field count for size of bookmarkCollection (maintaining of field will be implemented soon)

+++ src/main/java/com/neoxenus/webnovelreader/bookmarkcollection/dto/BookmarkCollectionDto.java
@@ -11,7 +11,7 @@ public record BookmarkCollectionDto(
         String name,
         String description,
         Integer position,
-        Long count
+        Integer count
 ) {
 
 }
